mobirace_be/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
sqlalchemy_utils
python-jose
mysql-connector-python
python-multipart



